feat: introduce partial classes for GoogleCloudPubSub
